Correct for API change.
Plex changed the way their API response was structured. This addresses that. It also adds a footer that includes the version number of the Plex Server it's checking against to aid in future troubleshooting. I've made this as unobtrusive as possible by making it gray, making it small font, and right-aligning it where nobody'll be looking, but if you want to exclude it, you can pass in the parameter "-OmitVersionNumber" and it'll skip that.